decopule PropertyFetchTypeResolver
